+++ frontend/components/forms/Login.tsx
@@ -15,7 +15,7 @@ const LoginForm: React.FC = () => {
     e.preventDefault()
 
     try {
-      const response = await fetch("http://127.0.0.1:8000/login", {
+      const response = await fetch("http://127.0.0.1:8000/user/login", {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
